Moved prototype-for to it's own file
